Remove OPTIONS methods from swagger
The purpose of options is to show the HTTP methods that the server supports for the route: https://www.w3schools.com/tags/ref_httpmethods.asp

Since you are generating a swagger, they don't need to be there.

Additionally, i think the operationId should be the route.name rather than the _handler.__name__ as if someone is using blueprints then there may be multiple routes that have the same operationId and that is against the swagger spec: http://swagger.io/specification/

Specifically this line:

operationId	string	Unique string used to identify the operation. The id MUST be unique among all operations described in the API. Tools and libraries MAY use the operationId to uniquely identify an operation, therefore, it is recommended to follow common programming naming conventions.